[ichthyop 0004054]: Application trackers should not be visible in the configuration file
(OutputManager.java)
Programatically added the application trackers : lon, lat, time, mortality and depth for 3D simulation.
Still read user-defined trackers in the configuration file.

(MortalityTracker.java)
Created application tracker that provides feedback about the state of the particle (whether at a given time it is dead or alive) and the cause of mortality.