Moved the templates in org.previmer.ichthyop.templates Created class Template.java that manages the templates.
Added wind action and drifter release file templates.

(TextFileEditor.java)
The default value of parameter type = textfile can be set as the name of a corresponding template file.
The template file has to exist in the template package.
So when user click new, the editor creates a copy of the template.