In Day360Calendar.java and InterannualCalendar.java deleted constructors that takes the origin of time as a String as it was generating problems. Updated the classes that used this constructor and replaced it by Calendar(year_o, month_o, day_o, hour_o, minute_o).
(TimeTracker.java)
Now systematically writes time variable attribute "origin" as the new 360_day calendar now takes it into account.

(WMSMapper.java)
Reads the time variable "origin" attribute no matter the type of the calendar. If the attribute does not exist, Ichthyop assumes the origin is 0001/01/01 00:00 as it used to be by default for the ClimatoCalendar.java